# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseListResponse"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: projectId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    put:
      operationId: updateDocument
      tags: [documents]
      summary: Update document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DASHBOARD ───
...
  /dashboard/monthly-financials:
    get:
      operationId: getMonthlyFinancials
      tags: [dashboard]
      summary: Monthly financials chart data
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Monthly financials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyFinancialsResponse"

  # ─── NOTIFICATIONS ───
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # EMPLOYEE
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        status:
          type: string
        projectId:
          type: integer
          nullable: true
        projectName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        submittedBy:
          type: string
        createdAt:
          type: string
      required: [id, title, category, amount, date, status, submittedBy, createdAt]
    CreateExpenseBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        projectId:
          type: integer
        description:
          type: string
        status:
          type: string
      required: [title, category, amount, date, status]
    UpdateExpenseBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        status:
          type: string
        description:
          type: string
    ExpenseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]
...
    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        projectId:
          type: integer
          nullable: true
        projectName:
          type: string
          nullable: true
        fileUrl:
          type: string
        fileType:
          type: string
        fileSize:
          type: string
        tags:
          type: array
          items:
            type: string
        uploadedBy:
          type: string
        createdAt:
          type: string
      required: [id, title, fileUrl, fileType, fileSize, tags, uploadedBy, createdAt]
    CreateDocumentBody:
      type: object
      properties:
        title:
          type: string
        projectId:
          type: integer
        fileUrl:
          type: string
        fileType:
          type: string
        fileSize:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title, fileUrl, fileType, fileSize]
    UpdateDocumentBody:
      type: object
      properties:
        title:
          type: string
        projectId:
          type: integer
        tags:
          type: array
          items:
            type: string
    DocumentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]

    # DASHBOARD
...
    MonthlyFinancialItem:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        expenses:
          type: number
        profit:
          type: number
      required: [month, revenue, expenses, profit]
    MonthlyFinancialsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyFinancialItem"
      required: [data]

    # NOTIFICATION
...